backend/requirements.txt
connexion[flask,swagger-ui]
flask
flask-sqlalchemy
pyodbc
pyjwt
passlib
argon2-cffi
swagger-ui-bundle
pydantic
